+++ src/tool/cpp/cppwinrt/strings/base_activation.h
@@ -429,10 +429,3 @@ WINRT_EXPORT namespace winrt
         return { result, take_ownership_from_abi };
     }
 }
-
-namespace std
-{
-    template<> struct hash<winrt::Windows::Foundation::IUnknown> : winrt::impl::hash_base<winrt::Windows::Foundation::IUnknown> {};
-    template<> struct hash<winrt::Windows::Foundation::IInspectable> : winrt::impl::hash_base<winrt::Windows::Foundation::IInspectable> {};
-    template<> struct hash<winrt::Windows::Foundation::IActivationFactory> : winrt::impl::hash_base<winrt::Windows::Foundation::IActivationFactory> {};
-}
